refactor: moved NavRoutes and NavItem into the presentation layer under navigation
