+++ app/src/main/java/com/oracle/visualize/presentation/navigation/NavItem.kt
@@ -1,7 +1,8 @@
-package com.oracle.visualize.domain.models
+package com.oracle.visualize.presentation.navigation
 
 import androidx.annotation.StringRes
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.oracle.visualize.presentation.navigation.NavRoutes
 
 /**
  * UI model representing a single item in the application's top-level navigation (e.g., Bottom Bar).
@@ -10,7 +11,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
  * and consistency across the app.
  *
  * @property label The string resource ID for the visible label of the navigation item.
- * @property icon The [ImageVector] displayed as the visual representation of the item.
+ * @property icon The [androidx.compose.ui.graphics.vector.ImageVector] displayed as the visual representation of the item.
  * @property badgeCount The count to display in an optional notification badge. Defaults to 0 (no badge).
  * @property destination The [NavRoutes] destination that this item targets.
  */
